[FRONT] Reorganize Utils folder.
