+++ nontclient/src/Utils/ComponentUtils/displayImage.js
@@ -1,0 +1,12 @@
+function arrayBufferToBase64(buffer) {
+  var binary = "";
+  var bytes = [].slice.call(new Uint8Array(buffer));
+  bytes.forEach((b) => (binary += String.fromCharCode(b)));
+  return window.btoa(binary);
+}
+
+function displayImage(buffer) {
+  arrayBufferToBase64(buffer);
+}
+
+export default displayImage;
